+++ MathVisualTests/Code/Game/TestShapeDisc.cpp
@@ -50,7 +50,7 @@ void TestShapeDisc::AddVertsForMe(std::vector<Vertex>& verts) const
 //-----------------------------------------------------------------------------------------------
 void TestShapeDisc::ChangeColor(Rgba8 newColor)
 {
-	for (int i = 0; i < m_lineVerts.size(); ++i)
+	for (int i = 0; i <  static_cast<int>(m_lineVerts.size()); ++i)
 	{
 		m_lineVerts[i].m_color = newColor;
 	}
