+++ MathVisualTests/Code/Game/TestShapeLine.cpp
@@ -52,7 +52,7 @@ void TestShapeLine::AddVertsForMe( std::vector<Vertex>& verts ) const
 //-----------------------------------------------------------------------------------------------
 void TestShapeLine::ChangeColor(Rgba8 newColor)
 {
-	for (int i = 0; i < m_lineVerts.size(); ++i)
+	for (int i = 0; i < static_cast<int>(m_lineVerts.size()); ++i)
 	{
 		m_lineVerts[i].m_color = newColor;
 	}
